[model] Make it possible to set the UUID during POSTs
@camillemarini it may be better to let compute forward you the
"model_end" UUID instead of forcing us to use a precomputed one. After
all, you don't know if train is going to succeed and if model_end will
ever be created. Besides, leaving the generation of the IDs of the
resources it stores to storage might be better in terms of separation
of concerns.

Maybe we could just send you this UUID (generated by storage) from
compute along with the performance result. What do you think ?